Matematiksel Model Destekli Atama Problemi
Determining the resource allocation of the main problem.
This issue is the creation of a mathematical model by making a literature review.
Using Gurobi solver to solve the mathematical model.
Visualization with python software with the results of this solver.
Using Excel as a database.
As a result;
Minimizing the appointment of the best people to these projects, taking into
account the availability of people who can work in software projects, the price of
personnel, project time, project competence, project priority, personnel
authority.
This project, in which these were visually seen with the gannt chart and the
densities of the personnel were visualized, was successfully completed and given
to the company.